rename key migration script and associated tests for clarity
